% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparecube.R
\name{add_changeval}
\alias{add_changeval}
\title{add_changeval}
\usage{
add_changeval(dt, val, by)
}
\arguments{
\item{dt}{data file ordered by}

\item{val}{value column to calculate change from}

\item{by}{dimensions to group the change value by}
}
\description{
Adds a change varible for a selected value columns. Change is defined as value/last value, with
last observation carried forward to get the change from last non-missing value within the strata.

Before calculating the change score, the data is sorted by the by columns and "AAR" to get the correct
order within each strata.

If < 2 unique years in the data file, nothing is done.
}
\keyword{internal}
