% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparecube.R
\name{add_outlier}
\alias{add_outlier}
\title{add_outlier}
\usage{
add_outlier(dt, val, by, change = FALSE)
}
\arguments{
\item{dt}{data file}

\item{val}{the value used for outlier detection}

\item{by}{All dimensions except AAR}

\item{change}{TRUE/FALSE, should outliers be detected based on the original or the change value}
}
\value{
dt
}
\description{
Add weighted quantiles and outlier information.

If change data is requested and there are < 2 unique years in the data file, nothing is done.
}
\keyword{internal}
