% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_cube}
\alias{aggregate_cube}
\title{aggregate_cube}
\usage{
aggregate_cube(cube, dim)
}
\arguments{
\item{cube}{cube file}

\item{dim}{dimension to aggregate on}
}
\description{
Aggregates value columns according to a given dimension variable.
If a total exists, the rows corresponding to the total is kept. If
no total exists, value columns will be aggregated according to their types.
Counts (TELLER, sumTELLER) are aggregated to the sum, while other value columns
are aggregated to their mean.
}
\keyword{internal}
