% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_cube_multi}
\alias{aggregate_cube_multi}
\title{aggregate_cube_multi}
\usage{
aggregate_cube_multi(cube, dimensions)
}
\arguments{
\item{cube}{data file}

\item{dimensions}{vector of dimensions to aggregate by}
}
\value{
data.table
}
\description{
Wrapper around \code{\link[=aggregate_cube]{aggregate_cube()}} to apply to multiple dimensions
#'
}
\keyword{internal}
