% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-helpers.R
\name{BoxPlot}
\alias{BoxPlot}
\title{BoxPlot}
\usage{
BoxPlot(
  data = dfnew_flag,
  onlynew = TRUE,
  change = FALSE,
  profileyear = PROFILEYEAR,
  data2 = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{data}{Dataset flagged for outliers}

\item{onlynew}{Should only new outliers be indicated on the plot? Default = TRUE}

\item{change}{Should plots be based on year-to-year changes. Default = FALSE}

\item{profileyear}{default = PROFILEYEAR}

\item{data2}{old file, only used}
}
\description{
Creates boxplots to visualize outliers.
Plots are stored in folders PLOTT/BP and PLOTT/BPc
}
