% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{check_befvekst}
\alias{check_befvekst}
\title{check_befvekst}
\usage{
check_befvekst(statusfolder = NULL)
}
\arguments{
\item{statusfolder}{name of statusfolder in access}
}
\description{
Loads BEFOLK_GK and BEFVEKST, as identified in the statusfolder in ACCESS.
BEFOLK_GK is filtered to contain the same strata as BEFVEKST, and population growth is calculated using lead TELLER values.
BEFVEKST TELLER is compared to the calculated population growth from BEFOLK, and all nonmatching rows, if any, are returned
}
