% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-censoring.R
\name{check_censoring}
\alias{check_censoring}
\title{check_censoring}
\usage{
check_censoring(dt)
}
\arguments{
\item{dt}{cube file}
}
\description{
Fetches censoring limits from ACCESS, and check if all values <= the censoring
limit has been removed. If ok, the function returns a confirmation.
If any observation below or equal to the limit is detected,
all rows containing unacceptable values are returned for inspection.
}
