% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-censoring.R
\name{check_coverage_geocode}
\alias{check_coverage_geocode}
\title{check_coverage_geocode}
\usage{
check_coverage_geocode(dt = newcube, save = TRUE)
}
\arguments{
\item{dt}{new data file}

\item{save}{save as csv?}
}
\value{
a table showing the proportion of censored rows per geo-code
}
\description{
For each geo-code, calculate the proportion of uncensored and censored rows per year.
}
