% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-censoring.R
\name{check_coverage_geolevel}
\alias{check_coverage_geolevel}
\title{check_coverage_geolevel}
\usage{
check_coverage_geolevel(dt = newcube, save = TRUE)
}
\arguments{
\item{dt}{new datafile}

\item{save}{save as csv?}
}
\description{
Calculates the number and proportion of GEO-codes per level getting at least 1 number.
}
