% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friskvik.R
\name{check_friskvik}
\alias{check_friskvik}
\title{check_friskvik}
\usage{
check_friskvik(
  profile = c("FHP", "OVP"),
  geolevel = c("B", "K", "F"),
  profileyear = NULL,
  test = FALSE
)
}
\arguments{
\item{profile}{"FHP" or "OVP"}

\item{geolevel}{"B", "K", or "F"}

\item{profileyear}{4-digit year}
}
\value{
csv-file in QualControl/FRISKVIKSJEKK
}
\description{
Runs a series of checks on the files in the most recent GODKJENT-folder, comparing the file to the corresponding
complete data file.
}
