% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungdata.R
\name{check_nevner_change}
\alias{check_nevner_change}
\title{check_nevner_change}
\usage{
check_nevner_change(dt = newcube, save = TRUE)
}
\arguments{
\item{dt}{cube file}
}
\description{
compare NEVNER towards the last and the maximum NEVNER in the time series.
Aggregate all dimensions except GEO and AAR
}
