% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungdata.R
\name{check_nevner_change_ungdata}
\alias{check_nevner_change_ungdata}
\title{check_nevner_change_ungdata}
\usage{
check_nevner_change_ungdata(dt = newcube)
}
\arguments{
\item{dt}{cube file}
}
\description{
compare NEVNER towards the last and the maximum NEVNER in the time series.
}
