% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-censoring.R
\name{compare_censoring}
\alias{compare_censoring}
\title{@title compare_censoring}
\usage{
compare_censoring(
  cube.new = newcube,
  cube.old = oldcube,
  by = NULL,
  filter.cubes = TRUE
)
}
\arguments{
\item{cube.new}{new KUBE file}

\item{cube.old}{old KUBE file or NULL}

\item{filter.cubes}{Should the cubes be filtered for better comparison? New and expired levels are removed,
and new/expired dimensions are aggregated. Default = TRUE}
}
\value{
a table containing the number of censored rows in the new and old KUBE, and the absolute and relative difference, grouped by type of SPVFLAGG and an additional dimension (optional)
}
\description{
Calculate number of censored observations and calculate absolute and relative difference
}
