% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-censoring.R
\name{compare_censoring_timeseries}
\alias{compare_censoring_timeseries}
\title{compare_censoring_timeseries}
\usage{
compare_censoring_timeseries(cube.new = newcube, cube.old = oldcube)
}
\arguments{
\item{cube.new}{new file}

\item{cube.old}{old file}
}
\description{
Compare censored observations within each time series. List number and proportion
of time series with 0, 1, 2, ..., n censored observations.
}
