% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-columns.R
\name{compare_colnames}
\alias{compare_colnames}
\title{CompareCols}
\usage{
compare_colnames(cube.new = newcube, cube.old = oldcube)
}
\arguments{
\item{data1}{New cube}

\item{data2}{Old cube or NULL}
}
\description{
Compare the columns across two cubes files
Report whether any new (not in old), or expired (not in new) columns are present.
}
