% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-columns.R
\name{compare_dimensions}
\alias{compare_dimensions}
\title{compare_dimensions}
\usage{
compare_dimensions(cube.new = NULL, cube.old = NULL)
}
\arguments{
\item{cube.new}{New cube}

\item{cube.old}{Old cube, or NULL}
}
\description{
Compares the unique levels of each dimension in the new file.
Checks if any levels are new (not in the old file) or expired (not in the new file).
}
