% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-geolevel.R
\name{compare_geolevels}
\alias{compare_geolevels}
\title{compare_geolevels}
\usage{
compare_geolevels(
  dt = newcube,
  comparison = c("FL", "LF", "KF", "FK", "BK", "KB", "OO")
)
}
\arguments{
\item{dt}{data file, defaults to newcube}

\item{comparison}{which geographical levels are compared. Accepts combinations "FL" (Fylke-Land),
"KF" (Kommune-Fylke"), "BK" (Bydel-Kommune), and "OO" (Oslo kommune-fylke).}
}
\value{
DT
}
\description{
Checks if a geographical level sums up to the next granularity.
}
\examples{
# compare_geolevels(data, "FL")
# compare_geolevels(data, "KF")
# compare_geolevels(data, "BK")
# compare_geolevels(data, "OO")
}
