% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-comparecube.R
\name{comparecube_summary}
\alias{comparecube_summary}
\title{comparecube_summary}
\usage{
comparecube_summary(dt = comparecube, save = TRUE)
}
\arguments{
\item{dt}{comparecube}
}
\value{
a DT output table
}
\description{
Generate a table estimating, for each strata by GEOniv/value column,
the number of identical/different rows, new/expired censored observations,
and mean/min/max diff / reldiff.

Uses \code{\link[=qc_round]{qc_round()}} on the data to eliminate very marginal differences.
}
