% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-comparecube.R
\name{diffvals_summary}
\alias{diffvals_summary}
\title{diffrow_summary}
\usage{
diffvals_summary(dt = comparecube, byyear = FALSE)
}
\arguments{
\item{dt}{comparecube}

\item{byyear}{Get output by year, default = FALSE}
}
\description{
Summarises number of diffrows and the sum of the diffs.
}
