% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-comparecube.R
\name{diffvals_which_levels}
\alias{diffvals_which_levels}
\title{diffvals_which_levels}
\usage{
diffvals_which_levels(dt = comparecube, dim = "GEO", val = "sumTELLER_uprikk")
}
\arguments{
\item{dt}{comparecube}

\item{dim}{dimension to list levels with diffs}

\item{val}{value column to identify diffs}
}
\description{
Prints a list of unique values of a dimensions with diffs of a specified value
}
