% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-helpers.R
\name{.findPlotSubset}
\alias{.findPlotSubset}
\title{Find subset of plots}
\usage{
.findPlotSubset(d, b, s)
}
\arguments{
\item{d}{dataset}

\item{b}{all bycols}

\item{s}{maximum number of panels per page}
}
\description{
The algorithm select the combination of dimensions yielding the maximum number of panels per page, but less than s
}
