% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-censoring.R
\name{explore_different_censoring}
\alias{explore_different_censoring}
\title{explore_different_censoring}
\usage{
explore_different_censoring(
  compare = comparecube,
  dt_new = newcube,
  dt_old = oldcube
)
}
\arguments{
\item{compare}{comparecube}

\item{dt_new}{newcube}

\item{dt_old}{oldcube}
}
\description{
filters out strata with different censoring and prints out censor-info columns to identify the reason for differences.
}
