% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_total}
\alias{find_total}
\title{find_total}
\usage{
find_total(cube, dim)
}
\arguments{
\item{cube}{datafile}

\item{dim}{dimension name}
}
\value{
the total category, or NA
}
\description{
Identifies if a total/aggregated level exist for a dimension. For ALDER,
the min_max is used as total if existing, for other dimensions 0 is the total value.
}
\examples{
# find_total(newcube, "ALDER")
}
\keyword{internal}
