% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparecube.R
\name{flag_outliers}
\alias{flag_outliers}
\title{flag_outliers}
\usage{
flag_outliers(cube, outlierval)
}
\arguments{
\item{cube}{data file}

\item{outlierval}{Which value is used to detect outliers? Selected with \code{\link[=select_outlier_pri]{select_outlier_pri()}}}
}
\value{
dt
}
\description{
Adds information related to outliers, defined as values outside the interval:
'weighted 25.percentile - 1.5IQR, weighted 75.percentile + 1.5IQR'

Also adds change value (if > 1 unique year) and change outliers.
}
