% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparecube.R
\name{flag_rows}
\alias{flag_rows}
\title{flag_rows}
\usage{
flag_rows(cube.new, cube.old = NULL, colinfo, flag = c("newrow", "exprow"))
}
\arguments{
\item{cube.new}{new data file}

\item{cube.old}{old data file}

\item{colinfo}{list generated with \link{identify_coltypes(cube.new, cube.old)}}

\item{flag}{one of "newrow" or "exprow", corresponding to which type of column should be flagged.}
}
\value{
flagged data file
}
\description{
Flags the new data file for new rows (not in the old file), and the old data file for expired rows (not in the new file).
}
\examples{
# flag_rows(cube.new, cube.old, colinfo, "newrow")
# flag_rows(cube.new, cube.old, colinfo, "exprow")
}
