% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_complete_strata}
\alias{get_complete_strata}
\title{get_complete_strata}
\usage{
get_complete_strata(
  data,
  by,
  type = c("censored", "missing"),
  valuecolumn = NULL
)
}
\arguments{
\item{data}{data file}

\item{by}{dimension columns to get complete strata by}

\item{type}{Should complete strata be identified by SPVFLAGG ("censored") or by missing value column ("missing")}

\item{valuecolumn}{must be provided when type = "missing", which value column should be used to identify complete strata?}
}
\description{
Returns only complete strata, as defined by the columns in \code{by}. Generate
a new column called \code{n_censored}, indicating the number of censored observations in
each strata, and returns rows where \code{n_censored} = 0.
\code{n_censored} is added by reference to the data file. To also filter the data, the data must be assigned, see examples.
}
\examples{
# Add n_censored, no filtering
# get_complete_strata(data, by = bycols, type = "censored")
# Actually filter data
# data <- get_complete_strata(data, by = bycols, type = "censored")
}
\keyword{internal}
