% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-utils.R
\name{get_duckdb_table}
\alias{get_duckdb_table}
\title{get_duckdb_table}
\usage{
get_duckdb_table(tablename)
}
\arguments{
\item{tablename}{name of table to read}
}
\value{
data.table
}
\description{
Fetches table from the local database into memory
}
