% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_coltypes}
\alias{identify_coltypes}
\title{identify_coltypes

uses internal object .validdims to identify dimension and value columns}
\usage{
identify_coltypes(cube.new = NULL, cube.old = NULL)
}
\arguments{
\item{cube.new}{new file}

\item{cube.old}{old file, or NULL}
}
\value{
a list containing information on dimension and value columns
}
\description{
identify_coltypes

uses internal object .validdims to identify dimension and value columns
}
