% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparecube.R
\name{make_comparecube}
\alias{make_comparecube}
\title{make_comparecube}
\usage{
make_comparecube(
  cube.new = NULL,
  cube.old = NULL,
  outliers = TRUE,
  dumps = NULL
)
}
\arguments{
\item{dumps}{}
}
\description{
make_comparecube
}
