% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-utils.R
\name{overwrite_duckdb_table}
\alias{overwrite_duckdb_table}
\title{overwrite_duckdb_table}
\usage{
overwrite_duckdb_table(d, tablename)
}
\arguments{
\item{d}{data}

\item{tablename}{table name}
}
\description{
Overwrites a table in the local duckdb. e.g. To be used if you have recoded av variable
}
