% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{plot_boxplot}
\usage{
plot_boxplot(dt = newcube_flag, onlynew = TRUE, change = FALSE, save = TRUE)
}
\arguments{
\item{dt}{flagged data file, default to newcube_flag}

\item{onlynew}{should only new outliers be plotted? Default = TRUE}

\item{change}{Should year-to-year changes be plotted? Default = FALSE}

\item{save}{Should plots be saved to the default folder? Default = T}
}
\description{
Generate boxplots indicating outliers (values outside q25-1.5IQR, q75+1.5IQR).
}
