% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-comparecube.R
\name{plot_diff_timetrends}
\alias{plot_diff_timetrends}
\title{plot_diff_timeseries}
\usage{
plot_diff_timetrends(dt = comparecube, save = TRUE)
}
\arguments{
\item{dt}{comparecube}

\item{save}{Should plots be saved in the Diff_timetrends folder? Default = TRUE. If FALSE, the plot is just printed in the console.}
}
\value{
plot
}
\description{
Plots differences between new and old cube file across time.
Used to see if there are any patterns in the diffs over time. For indicators with strong trends,
the differences are expected to become smaller with time as the data is standardized to the last year/period
}
