% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_bydel.R
\name{plot_timeseries_bydel}
\alias{plot_timeseries_bydel}
\title{plot_timeseries_bydel}
\usage{
plot_timeseries_bydel(dt = newcube_flag, save = TRUE)
}
\arguments{
\item{dt}{flagged data file, default to newcube_flag}
}
\description{
Generate timeseries plots for bydel, with weighted
}
