% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_country.R
\name{plot_timeseries_country}
\alias{plot_timeseries_country}
\title{plot_timeseries_country}
\usage{
plot_timeseries_country(cube, save = TRUE)
}
\arguments{
\item{cube}{data file}

\item{save}{Should the plot be saved in the TimeSeries_country folder? Default = TRUE.
If FALSE, the plot is just printed in the console.}
}
\value{
plot
}
\description{
Plots country-level time series across all dimensions except GEO and AAR.
Useful to detect classification errors or other time series anomalies.
}
