% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcontrol-package.R
\docType{package}
\name{qualcontrol-package}
\alias{qualcontrol}
\alias{qualcontrol-package}
\title{qualcontrol: Quality Control}
\description{
Quality control of publication-ready files.
}
\author{
\strong{Maintainer}: Vegard Lysne \email{vegard.lysne@helsedir.no} (\href{https://orcid.org/0000-0002-0816-5075}{ORCID})

Authors:
\itemize{
  \item Vegard Lysne \email{vegard.lysne@helsedir.no} (\href{https://orcid.org/0000-0002-0816-5075}{ORCID})
}

}
\keyword{internal}
