% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-helpers.R
\name{TimelineBydel}
\alias{TimelineBydel}
\title{TimelineBydel}
\usage{
TimelineBydel(data = dfnew_flag, profileyear = PROFILEYEAR, overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\description{
Plots timelines for each bydel, with the weighted mean timeline for the bydel and kommune superimposed.
Used to check the validity of data on bydel level. The mean timeline for bydel should approximately match
the timeline for kommune.
}
