% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-tabs.R
\name{update_georecode}
\alias{update_georecode}
\title{update_georecode}
\usage{
update_georecode(year, overwrite = T)
}
\arguments{
\item{year}{the year for valid geo codes}

\item{overwrite}{should the current table be replaced by the new?}
}
\description{
Update the georecode table used for recoding geographical codes to current codes for comparison.
Connects to the geo-code database and use information for kommune, fylke.
}
\examples{
# qualcontrol:::update_georecode(2024, overwrite = T)
}
\concept{family name}
