% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_qcyear}
\alias{update_qcyear}
\title{update_qcyear}
\usage{
update_qcyear(year = NULL)
}
\arguments{
\item{year}{corresponding to which output folder you want the output to go into.
Default value is given in the config_qualcontrol.yml file.}
}
\description{
update_qcyear
}
\examples{
# To use 2025 as the base folder for output:
# update_qcyear(year = 2025)
# To use the year as defined in the config file:
# update_qcyear(year = NULL)
}
