% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friskvik.R
\name{check_friskvik}
\alias{check_friskvik}
\title{check_friskvik}
\usage{
check_friskvik(
  profile = c("FHP", "OVP"),
  geolevel = c("B", "K", "F"),
  profileyear = NULL,
  save = TRUE,
  test = FALSE
)
}
\arguments{
\item{profile}{"FHP" or "OVP"}

\item{geolevel}{"B", "K", or "F"}

\item{profileyear}{4-digit year}

\item{save}{write CSV?}

\item{test}{write to testfolder?}
}
\value{
Table in global environment and csv-file in QualControl/FRISKVIKSJEKK if save = TRUE
}
\description{
Runs a series of checks on the files in the most recent GODKJENT-folder, comparing the file to the corresponding
complete data file.
}
