% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_qcfolders}
\alias{generate_qcfolders}
\title{generate_folders}
\usage{
generate_qcfolders(cubename, year = NULL)
}
\arguments{
\item{cubename}{name of cube}

\item{year}{profileyear, defaults to options(qualcontrol.year)}
}
\description{
Create Folder structure according to profile year and cube name of new data file
}
\keyword{internal}
