% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_qcfolders}
\alias{generate_qcfolders}
\title{generate_folders

Create Folder structure according to profile year and cube name of new data file}
\usage{
generate_qcfolders(year, cubename)
}
\description{
generate_folders

Create Folder structure according to profile year and cube name of new data file
}
