% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-utils.R
\name{get_duckdb_table}
\alias{get_duckdb_table}
\title{get_duckdb_table}
\usage{
get_duckdb_table(tablename)
}
\arguments{
\item{tablename}{navn på tabellen du vil hente}
}
\description{
Henter en tabell fra den lokale duckdb-databasen inn i minnet
Nyttig dersom du trenger å se en tabell, eller kode om en variabel
(f.eks. for å matche oldcube mot newcube), før overskriving med \code{overwrite_duckdb_table}
}
\seealso{
Other duck:
\code{\link[=overwrite_duckdb_table]{overwrite_duckdb_table()}}
}
\concept{duck}
