% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparecube.R
\name{make_comparecube}
\alias{make_comparecube}
\title{make_comparecube}
\usage{
make_comparecube(
  cube.new = newcube,
  cube.old = oldcube,
  outliers = TRUE,
  dumps = getOption("qualcontrol.dumps")
)
}
\arguments{
\item{cube.new}{new file, default = newcube}

\item{cube.old}{old file, default = oldcube. Can be NULL.}

\item{outliers}{should outliers be flagged using \code{\link[=flag_outliers]{flag_outliers()}}?}

\item{dumps}{which files to save, defaults to getOption("qualcontrol.dumps"). For no dumps, set to NULL.}

\item{overwrite}{should existing files (exact filename) be overwritten?}
}
\value{
flagged cube objects, comparecube, and saved csv-files if requested.
}
\description{
make_comparecube
}
