% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-utils.R
\name{overwrite_duckdb_table}
\alias{overwrite_duckdb_table}
\title{overwrite_duckdb_table}
\usage{
overwrite_duckdb_table(d, tablename)
}
\arguments{
\item{tablename}{navn på tabellen du vil hente}
}
\description{
Overskriver en tabell i den lokale duckdb-databasen.
Nyttig om du har omkodet en variabel for å matche oldcube mot newcube.
}
\seealso{
Other duck:
\code{\link[=get_duckdb_table]{get_duckdb_table()}}
}
\concept{duck}
