% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_country.R
\name{plot_timeseries_country}
\alias{plot_timeseries_country}
\title{plot_timeseries_country}
\usage{
plot_timeseries_country(cube, save = TRUE)
}
\arguments{
\item{cube}{data file}
}
\value{
list
}
\description{
Plots country-level time series across all dimensions except GEO and AAR.
Useful to detect classification errors or other time series anomalies.
}
