% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-files.R
\name{readfiles}
\alias{readfiles}
\title{readcubes}
\usage{
readfiles(
  cube.new = NULL,
  recode.new = FALSE,
  cube.old = NULL,
  recode.old = FALSE,
  comparecube = TRUE,
  outliers = TRUE,
  dumps = getOption("qualcontrol.dumps")
)
}
\arguments{
\item{cube.new}{Full name of new cube file, including date-tag}

\item{recode.new}{TRUE/FALSE, recode GEO-codes in new cube?}

\item{cube.old}{Full name of new cube file, including date-tag}

\item{recode.old}{TRUE/FALSE, recode GEO-codes in old cube?}

\item{comparecube}{TRUE/FALSE, should comparecube be generated (default = TRUE)}

\item{outliers}{Should outliers be detected and flagged?}

\item{dumps}{save csv-files? Available options are \code{comparecube}, \code{newcube_flag}, and \code{oldcube_flag}}
}
\value{
New and old file with attributes
}
\description{
Reads cube files for quality control routines
}
