% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-files.R
\name{readfiles}
\alias{readfiles}
\title{readcubes}
\usage{
readfiles(
  cube.new = NULL,
  modus.new = c("KH", "NH"),
  recode.new = FALSE,
  cube.old = NULL,
  modus.old = c("KH", "NH"),
  recode.old = FALSE
)
}
\arguments{
\item{newcubefile}{new cube filename, in the format "NAME_1234-12-34-56-78"}

\item{modusnew}{"KH" or "NH"}

\item{recode.newcube}{TRUE/FALSE, should GEO codes in new file be recoded to current GEO}

\item{oldcubefile}{old/comparison cube filename, in the format "NAME_1234-12-34-56-78" or NULL}

\item{modusold}{"KH" or "NH"}

\item{recode.oldcube}{TRUE/FALSE, should GEO codes in old file be recoded to current GEO}
}
\value{
New and old file with attributes
}
\description{
Reads cube files for quality control routines
}
