% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-geolevel.R
\name{unknown_bydel}
\alias{unknown_bydel}
\title{unknown_bydel}
\usage{
unknown_bydel(cube)
}
\arguments{
\item{cube}{data file}
}
\value{
DT
}
\description{
Estimate the proportion of unknown bydel, to check validity of the data.
}
