% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-geolevel.R
\name{unknown_bydel}
\alias{unknown_bydel}
\title{unknown_bydel}
\usage{
unknown_bydel(cube, maxrows = TRUE)
}
\arguments{
\item{cube}{data file}

\item{maxrows}{Should the output table ble cropped to show a maximum of 4000 observations?}
}
\value{
DT
}
\description{
Estimate the proportion of unknown bydel, to check validity of the data.
}
