% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-geolevel.R
\name{unknown_bydel}
\alias{unknown_bydel}
\title{unknown_bydel}
\usage{
unknown_bydel(dt = newcube, crop = TRUE, maxrows = 4000, save = TRUE)
}
\arguments{
\item{dt}{data file, defaults to newcube}

\item{crop}{Should the output table ble cropped to show a maximum of \code{maxrows}observations?}

\item{maxrows}{Max observations to show if \code{crop} = TRUE. Default = 4000. Will be divided equally between strata}
}
\value{
DT
}
\description{
Estimate the proportion of unknown bydel, to check validity of the data.
}
