% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-tabs.R
\name{update_internal_data}
\alias{update_internal_data}
\title{Update internal data
Stores objects in R/sysdata.rda which is needed in the package}
\usage{
update_internal_data(year)
}
\arguments{
\item{year}{year for validgeo}
}
\value{
\itemize{
\item .validdims
\item .validgeo
}
}
\description{
Update internal data
Stores objects in R/sysdata.rda which is needed in the package
}
