% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-tabs.R
\name{update_internal_data}
\alias{update_internal_data}
\title{Update internal data}
\usage{
update_internal_data(geoyear, overwrite = TRUE)
}
\arguments{
\item{geoyear}{year for valid GEO-codes}

\item{overwrite}{overwrite data?}
}
\value{
\itemize{
\item .standarddims
\item .validdims
\item .validgeo
\item .popinfo
\item .georecode
}
}
\description{
Stores objects in R/sysdata.rda which is needed in the package
}
\keyword{internal}
