% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-tabs.R
\name{update_popinfo}
\alias{update_popinfo}
\title{update_popinfo}
\usage{
update_popinfo(popfile, overwrite = TRUE)
}
\arguments{
\item{popfile}{complete file name of BEFOLK_GK file}

\item{overwrite}{should the file be overwritten?}

\item{year}{referring to NESSTAR-folder}
}
\description{
Generates popinfo.rds file containing weights and GEOniv including small/large kommune and Helseregion
}
\examples{
# update_popinfo("O:/Prosjekt/FHP/PRODUKSJON/PRODUKTER/KUBER/STATBANK/STATBANK_2026/BEFOLK_GK_2025-08-07-12-50.csv",
overwrite = TRUE)

}
