% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-tabs.R
\name{update_popinfo}
\alias{update_popinfo}
\title{update_popinfo}
\usage{
update_popinfo(popfile, overwrite = T)
}
\arguments{
\item{popfile}{complete file name of BEFOLK_GK file}

\item{overwrite}{should the file be overwritten?}

\item{year}{referring to NESSTAR-folder}
}
\description{
Generates popinfo.rds file containing weights and GEOniv including small/large kommune and Helseregion
}
